feat: rename InitialPrompt to BusinessInfoPrompt, modify prompt and check for valid url
